refactor(server): move Docker-specific modules into services/docker/
- Move docker_diagnostics, port_allocator, ossfs_mixin into docker/
- Move docker_snapshot_runtime into docker/
- Split docker_windows_profile into docker/windows_profile (Docker-specific)
  and windows_profile_common (shared with k8s)
- Update all import paths across codebase and tests
