+++ server/tests/test_docker_diagnostics.py
@@ -15,7 +15,7 @@
 from types import SimpleNamespace
 from unittest.mock import MagicMock, patch
 
-from opensandbox_server.services.docker_diagnostics import (
+from opensandbox_server.services.docker.docker_diagnostics import (
     DockerDiagnosticsMixin,
     _parse_since_to_timestamp,
 )
@@ -40,7 +40,7 @@ def _container(attrs: dict):
 
 
 def test_parse_since_to_timestamp_uses_valid_units_and_default() -> None:
-    with patch("opensandbox_server.services.docker_diagnostics.time.time", return_value=1000):
+    with patch("opensandbox_server.services.docker.docker_diagnostics.time.time", return_value=1000):
         assert _parse_since_to_timestamp("2m") == 880
         assert _parse_since_to_timestamp("1 h") == -2600
         assert _parse_since_to_timestamp("nonsense") == 400
@@ -51,7 +51,7 @@ def test_get_sandbox_logs_decodes_bytes_and_passes_since() -> None:
     service = _DiagnosticsService(container)
 
     with patch(
-        "opensandbox_server.services.docker_diagnostics._parse_since_to_timestamp",
+        "opensandbox_server.services.docker.docker_diagnostics._parse_since_to_timestamp",
         return_value=123,
     ):
         result = service.get_sandbox_logs("sbx-1", tail=25, since="5m")
