+++ server/tests/test_docker_service.py
@@ -2367,10 +2367,10 @@ class TestDockerVolumeValidation:
             ],
         )
 
-        with patch("opensandbox_server.services.ossfs_mixin.os.name", "posix"):
-            with patch("opensandbox_server.services.ossfs_mixin.os.path.ismount", return_value=False):
-                with patch("opensandbox_server.services.ossfs_mixin.os.makedirs"):
-                    with patch("opensandbox_server.services.ossfs_mixin.subprocess.run") as mock_run:
+        with patch("opensandbox_server.services.docker.ossfs_mixin.os.name", "posix"):
+            with patch("opensandbox_server.services.docker.ossfs_mixin.os.path.ismount", return_value=False):
+                with patch("opensandbox_server.services.docker.ossfs_mixin.os.makedirs"):
+                    with patch("opensandbox_server.services.docker.ossfs_mixin.subprocess.run") as mock_run:
                         mock_run.return_value = MagicMock(returncode=1, stderr="mount failed")
                         with pytest.raises(HTTPException) as exc_info:
                             await service.create_sandbox(request)
@@ -2395,7 +2395,7 @@ class TestDockerVolumeValidation:
             mount_path="/mnt/data",
         )
 
-        with patch("opensandbox_server.services.ossfs_mixin.os.name", "nt"):
+        with patch("opensandbox_server.services.docker.ossfs_mixin.os.name", "nt"):
             with pytest.raises(HTTPException) as exc_info:
                 service._validate_ossfs_volume(volume)
         assert exc_info.value.status_code == status.HTTP_400_BAD_REQUEST
@@ -2422,8 +2422,8 @@ class TestDockerVolumeValidation:
         )
         backend_path = "/mnt/ossfs/bucket-test-3/task-001"
 
-        with patch("opensandbox_server.services.ossfs_mixin.os.makedirs"):
-            with patch("opensandbox_server.services.ossfs_mixin.subprocess.run") as mock_run:
+        with patch("opensandbox_server.services.docker.ossfs_mixin.os.makedirs"):
+            with patch("opensandbox_server.services.docker.ossfs_mixin.subprocess.run") as mock_run:
                 mock_run.return_value = MagicMock(returncode=0, stderr="")
                 service._mount_ossfs_backend_path(volume, backend_path)
 
@@ -2457,8 +2457,8 @@ class TestDockerVolumeValidation:
         )
         backend_path = "/mnt/ossfs/bucket-test-3/task-001"
 
-        with patch("opensandbox_server.services.ossfs_mixin.os.makedirs"):
-            with patch("opensandbox_server.services.ossfs_mixin.subprocess.run") as mock_run:
+        with patch("opensandbox_server.services.docker.ossfs_mixin.os.makedirs"):
+            with patch("opensandbox_server.services.docker.ossfs_mixin.subprocess.run") as mock_run:
                 mock_run.return_value = MagicMock(returncode=0, stderr="")
                 service._mount_ossfs_backend_path(volume, backend_path)
 
@@ -2536,10 +2536,10 @@ class TestDockerVolumeValidation:
             ],
         )
 
-        with patch("opensandbox_server.services.ossfs_mixin.os.name", "posix"):
-            with patch("opensandbox_server.services.ossfs_mixin.os.path.ismount", return_value=False):
-                with patch("opensandbox_server.services.ossfs_mixin.os.makedirs"):
-                    with patch("opensandbox_server.services.ossfs_mixin.subprocess.run") as mock_run:
+        with patch("opensandbox_server.services.docker.ossfs_mixin.os.name", "posix"):
+            with patch("opensandbox_server.services.docker.ossfs_mixin.os.path.ismount", return_value=False):
+                with patch("opensandbox_server.services.docker.ossfs_mixin.os.makedirs"):
+                    with patch("opensandbox_server.services.docker.ossfs_mixin.subprocess.run") as mock_run:
                         mock_run.return_value = MagicMock(returncode=0, stderr="")
                         with patch.object(service, "_ensure_image_available"), patch.object(
                             service, "_prepare_sandbox_runtime"
@@ -2585,9 +2585,9 @@ class TestDockerVolumeValidation:
             ),
         ]
 
-        with patch("opensandbox_server.services.ossfs_mixin.os.path.ismount", return_value=False):
-            with patch("opensandbox_server.services.ossfs_mixin.os.makedirs"):
-                with patch("opensandbox_server.services.ossfs_mixin.subprocess.run") as mock_run:
+        with patch("opensandbox_server.services.docker.ossfs_mixin.os.path.ismount", return_value=False):
+            with patch("opensandbox_server.services.docker.ossfs_mixin.os.makedirs"):
+                with patch("opensandbox_server.services.docker.ossfs_mixin.subprocess.run") as mock_run:
                     mock_run.return_value = MagicMock(returncode=0, stderr="")
                     mount_keys = service._prepare_ossfs_mounts(volumes)
 
@@ -2659,8 +2659,8 @@ class TestDockerVolumeValidation:
         service = DockerSandboxService(config=_app_config())
         service._ossfs_mount_ref_counts[mount_key] = 1
 
-        with patch("opensandbox_server.services.ossfs_mixin.os.path.ismount", return_value=True):
-            with patch("opensandbox_server.services.ossfs_mixin.subprocess.run") as mock_run:
+        with patch("opensandbox_server.services.docker.ossfs_mixin.os.path.ismount", return_value=True):
+            with patch("opensandbox_server.services.docker.ossfs_mixin.subprocess.run") as mock_run:
                 mock_run.return_value = MagicMock(returncode=0, stderr="")
                 service.delete_sandbox("sandbox-1")
 
@@ -2673,8 +2673,8 @@ class TestDockerVolumeValidation:
         mock_docker.from_env.return_value = MagicMock()
         service = DockerSandboxService(config=_app_config())
 
-        with patch("opensandbox_server.services.ossfs_mixin.os.path.ismount", return_value=True):
-            with patch("opensandbox_server.services.ossfs_mixin.subprocess.run") as mock_run:
+        with patch("opensandbox_server.services.docker.ossfs_mixin.os.path.ismount", return_value=True):
+            with patch("opensandbox_server.services.docker.ossfs_mixin.subprocess.run") as mock_run:
                 service._release_ossfs_mount(mount_key)
 
         mock_run.assert_not_called()
@@ -2736,8 +2736,8 @@ class TestDockerVolumeValidation:
         service = DockerSandboxService(config=_app_config())
         assert service._ossfs_mount_ref_counts[mount_key] == 2
 
-        with patch("opensandbox_server.services.ossfs_mixin.os.path.ismount", return_value=True):
-            with patch("opensandbox_server.services.ossfs_mixin.subprocess.run") as mock_run:
+        with patch("opensandbox_server.services.docker.ossfs_mixin.os.path.ismount", return_value=True):
+            with patch("opensandbox_server.services.docker.ossfs_mixin.subprocess.run") as mock_run:
                 service.delete_sandbox("sandbox-a")
 
         assert service._ossfs_mount_ref_counts[mount_key] == 1
